# api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FutLendas API",
    "version": "1.0.0",
    "description": "Documentação da API backend para o aplicativo FutLendas."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor de Desenvolvimento Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "GoleiroStats": {
        "type": "object",
        "description": "Schema com as estatísticas de desempenho de um goleiro.",
        "properties": {
          "jogador_id": {
            "type": "integer",
            "description": "ID do jogador (goleiro).",
            "example": 5
          },
          "nome_goleiro": {
            "type": "string",
            "description": "Nome do goleiro.",
            "example": "Alisson Becker"
          },
          "total_jogos": {
            "type": "integer",
            "description": "Número total de partidas jogadas como goleiro.",
            "example": 10
          },
          "total_gols_sofridos": {
            "type": "integer",
            "description": "Número total de gols sofridos.",
            "example": 8
          },
          "total_vitorias": {
            "type": "integer",
            "description": "Número de partidas vencidas como goleiro.",
            "example": 6
          },
          "total_derrotas": {
            "type": "integer",
            "description": "Número de partidas perdidas como goleiro.",
            "example": 2
          },
          "total_segundos_jogados": {
            "type": "integer",
            "description": "Tempo total em segundos jogado como goleiro.",
            "example": 5400
          },
          "total_minutos_jogados": {
            "type": "number",
            "format": "float",
            "description": "Tempo total em minutos jogado como goleiro.",
            "example": 90
          },
          "media_gols_sofridos_por_minuto": {
            "type": "number",
            "format": "float",
            "description": "Média de gols sofridos por minuto em campo. Usado para ranking.",
            "example": 0.09
          }
        }
      },
...
      "Liga": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nome": { "type": "string", "example": "Temporada 2025/1" },
          "data_inicio": { "type": "string", "format": "date", "example": "2025-01-01" },
          "data_fim": { "type": "string", "format": "date", "example": "2025-06-30" },
          "finalizada_em": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Data e hora em que a liga foi efetivamente finalizada. Nulo se ainda estiver ativa.",
            "example": "2025-06-15T10:30:00.000Z"
          }
        }
      },
      "Rodada": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "liga_id": { "type": "integer", "example": 1 },
          "data": { "type": "string", "format": "date", "example": "2025-01-15" },
          "status": { "type": "string", "enum": ["aberta", "finalizada"], "example": "aberta" }
        }
      },
...
      "RodadaInput": {
        "type": "object",
        "required": ["data"],
        "properties": { "data": { "type": "string", "format": "date", "example": "2025-07-15" } }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/goleiro-stats/liga": {
      "get": {
        "summary": "Estatísticas dos goleiros por Liga",
        "description": "Retorna um ranking de desempenho de todos os goleiros que jogaram em uma liga específica, ordenado pela melhor média de gols sofridos.",
        "tags": ["Estatísticas de Goleiros"],
        "parameters": [
          {
            "in": "query",
            "name": "liga_id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "O ID da liga para a qual as estatísticas serão calculadas.",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de estatísticas dos goleiros retornado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/GoleiroStats" }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida: o query parameter 'liga_id' é obrigatório."
          }
        }
      }
    },
...
    "/ligas": {
      "post": {
        "summary": "Criar nova liga (Admin)",
        "description": "Cria uma nova liga ou temporada no sistema.",
        "tags": ["Ligas"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "data_inicio", "data_fim"],
                "properties": {
                  "nome": {
                    "type": "string",
                    "description": "Nome da liga.",
                    "example": "Liga de Verão 2025"
                  },
                  "data_inicio": {
                    "type": "string",
                    "format": "date",
                    "description": "Data de início da liga (YYYY-MM-DD).",
                    "example": "2025-01-01"
                  },
                  "data_fim": {
                    "type": "string",
                    "format": "date",
                    "description": "Data de término da liga (YYYY-MM-DD).",
                    "example": "2025-06-30"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Liga criada com sucesso." },
          "400": { "description": "Dados inválidos na requisição." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado (não é admin)." }
        }
      },
      "get": {
        "summary": "Listar todas as ligas",
        "description": "Retorna a lista completa de ligas cadastradas, incluindo as finalizadas.",
        "tags": ["Ligas"],
        "responses": {
          "200": {
            "description": "Lista de ligas retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Liga" } }
              }
            }
          }
        }
      }
    },
    "/ligas/{id}": {
      "get": {
        "summary": "Buscar liga por ID",
        "description": "Retorna os detalhes de uma liga específica.",
        "tags": ["Ligas"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "ID da liga.",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da liga.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Liga" } } }
          },
          "404": { "description": "Liga não encontrada." }
        }
      },
      "put": {
        "summary": "Atualizar liga (Admin)",
        "description": "Atualiza informações de uma liga existente.",
        "tags": ["Ligas"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID da liga."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string", "example": "Liga de Inverno 2025" },
                  "data_inicio": { "type": "string", "format": "date", "example": "2025-07-01" },
                  "data_fim": { "type": "string", "format": "date", "example": "2025-12-31" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Liga atualizada com sucesso." },
          "400": { "description": "Dados inválidos." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Liga não encontrada." }
        }
      },
      "delete": {
        "summary": "Deletar liga (Admin)",
        "description": "Remove permanentemente uma liga do sistema. **Atenção:** Esta ação é destrutiva e usará o `ON DELETE CASCADE` do banco de dados para remover todas as rodadas, partidas e estatísticas associadas a esta liga.\n",
        "tags": ["Ligas"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID da liga a ser deletada."
          }
        ],
        "responses": {
          "200": { "description": "Liga deletada com sucesso." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Liga não encontrada." }
        }
      }
    },
    "/ligas/{id}/finalizar": {
      "post": {
        "summary": "Finalizar liga antecipadamente (Admin)",
        "description": "Marca a liga como finalizada na data e hora atuais. Uma liga finalizada não pode ter novas rodadas criadas. Esta ação é irreversível pelo sistema.\n",
        "tags": ["Ligas"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID da liga a ser finalizada."
          }
        ],
        "responses": {
          "200": {
            "description": "Liga finalizada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Liga finalizada com sucesso!" },
                    "finalizada_em": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-11-16T20:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "A liga já foi finalizada anteriormente." },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Liga não encontrada." }
        }
      }
    },
...
    "/rodadas/liga/{liga_id}": {
      "get": {
        "summary": "Listar rodadas de uma liga",
        "description": "Retorna todas as rodadas (passadas e futuras) de uma liga específica, ordenadas por data.",
        "tags": ["Rodadas"],
        "parameters": [
          { "in": "path", "name": "liga_id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de rodadas retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Rodada" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar nova rodada (Admin)",
        "description": "Cria uma nova rodada dentro de uma liga, desde que a liga não esteja finalizada e a data seja válida.",
        "tags": ["Rodadas"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "liga_id", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RodadaInput" } }
          }
        },
        "responses": {
          "201": { "description": "Rodada criada com sucesso." },
          "400": {
            "description": "Dados inválidos, ou a liga já está finalizada, ou a data da rodada é posterior ao fim da liga."
          },
          "401": { "description": "Não autenticado." },
          "403": { "description": "Acesso negado." },
          "404": { "description": "Liga não encontrada." },
          "409": { "description": "Conflito: Já existe uma rodada nesta data para esta liga." }
        }
      }
    },
...
  },
...
}